move everything related to grammar out of qgis loader
